requirements.txt
asgiref==3.9.2
Django==4.2.24
django-admin==2.0.2
django-cors-headers==4.9.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.1
excel-base==1.0.4
isoweek==1.3.3
joblib==1.5.2
numpy==2.0.2
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
screen==1.0.1
six==1.17.0
sklearn==0.0
sqlparse==0.5.3
threadpoolctl==3.6.0
TimeConvert==3.0.13
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
xlwt==1.3.0
